chat component cleanup
